pyproject.toml
[tool.poetry]
name = "christmas"
version = "0.1.0"
description = ""
authors = ["sumin <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
supabase = "^2.0.0"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
